requirements.txt
flask==2.3.2
matplotlib==3.7.2
Pillow==10.0.0
tensorflow==2.13.0